requirements.txt
# Telegram bot
python-telegram-bot==21.5

# Google Sheets
gspread==6.1.2
google-auth==2.29.0

# OCR (photos & scanned PDFs)
pytesseract==0.3.13
Pillow==10.4.0

# PDF parsing
pdfplumber==0.11.4

# Env vars
python-dotenv==1.0.1
